ENH: add partial_derivative method to bivariate spline classes.
A new method, "partial_derivative", is added to the base class for
bivariate splines in scipy.interpolate. The new method is able to create
a new spline instance that, when called, evaluates the indicated partial
derivative of the original spline instance.

The method complements the existing "__call__" method, and is analogous
to the "derivative" method of univariate splines. Unlike the existing
"__call__" or "ev" methods, "partial_derivative" does not re-calculate
the new spline coefficients for the derivative each time it is called.