MAINT: refactor derivative spline object's inheritance.
A new class, "_DerivedBivariateSpline", is created as the class of
bivariate splines constructed from a transformation on original spline's
knots and coefficients directly. The class inherits from BivariateSpline
and hence support spline evaluation methods. However, the "fp" and
"get_residual" attributes are overridden, because its construction
refers to no interpolated data.

The class "_DerivedBivariateSpline" is not supposed to be instantiated
in the typical manner in the client code. It is excluded from the
"__all__" attribute of "fitpack2", and is not imported by the
"__init__" module of the "scipy.interpolate" package.