MAINT: Refactor extensions so that C++ sources are not mixed with C.
The functions in sf_error.c is used by C++ extensions as well. However,
controlling the compiler flags can be tricky for mixed-language sources.

After refactoring, sf_error.c is compiled separately as a C library, and
the C++ extension is compiled from C++ sources only. The latter is then
linked with the sf_error library. In this way we avoid mixed-language
source compiling.